[feat] - config elasticsearch to projecT

+++ src/main/java/ck4/nvb/rsmanagement/core/module/stores/product/domain/ProductRepository.java
@@ -18,5 +18,5 @@ public interface ProductRepository extends BaseFullAuditedRepository<Product, Lo
       nativeQuery = true)
   int remainQuantity(Long productId, Long storeId);
 
-  int countProductsByDeletedIsFalse();
+  long countProductsByDeletedIsFalse();
 }
